Unify search fields
